third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/1.1/scripthash-handler-allowed-expected.txt

CONSOLE ERROR: line 22: Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src 'sha256-nhtYaXCssBJTThiDLYewspQYue9tisulEwJ3nTJKcMI=' 'unsafe-hashed-attributes'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

This is a testharness.js-based test.
PASS Inline event handlers whitelisted by the policy should fire. 
PASS Inline event handlers not whitelisted by the policy should generate error events. 
Harness: the test ran to completion.
